fix: Critical database type mismatch and production safety improvements
- Fixed f32/f64 type mismatch between Rust code and PostgreSQL schema that prevented all memory storage operations
- Replaced all unwrap() and expect() calls with proper error handling in main.rs and mcp/server.rs
- Fixed signal handler error handling to prevent panics on startup
- Applied code formatting and resolved linting issues
- Improved error propagation in MCP server JSON serialization

This resolves the critical production bug where importance_score type mismatch caused all database operations to fail.

+++ src/memory/repository.rs
@@ -437,7 +437,7 @@ impl MemoryRepository {
                 Some(ScoreExplanation {
                     semantic_contribution: similarity_score * 0.4,
                     temporal_contribution: temporal_score.unwrap_or(0.0) * 0.3,
-                    importance_contribution: importance_score * 0.2,
+                    importance_contribution: (importance_score * 0.2) as f32,
                     access_frequency_contribution: access_frequency_score.unwrap_or(0.0) * 0.1,
                     total_score: combined_score,
                     factors: vec![
